cinema_backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-allauth
dj-rest-auth
django-cors-headers
django-filter
python-dotenv
psycopg2-binary
requests
Pillow